updated files

+++ kernel.h
@@ -1,5 +1,5 @@
 double dot_product(unsigned int n, double VectorA[], double VectorB[]) {
-    long double sdot = 0.0f;
+    double sdot = 0.0f;
     unsigned int i;
 
     for (i = 0; i < n; i++)
